% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind_fill.R
\name{cbind_fill}
\alias{cbind_fill}
\title{cbind function for dataframes and arrays that works with one empty input}
\usage{
cbind_fill(...)
}
\arguments{
\item{...}{dataframes or arrays to be column binded.}
}
\value{
a dataframe with your data
}
\description{
cbind function for dataframes and arrays that works with one empty input
}
\examples{
t1 <- data.frame()
t2 <- c(1,2,3,4,5)
out <- cbind_fill(t1,t2)


t1 <- data.frame()
t2 <- mtcars
out <- cbind_fill(t1,t2)

}
