% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip.R
\name{clip}
\alias{clip}
\title{Copy dataframe to clipboard for paste into Excel}
\usage{
clip(x, row.names = FALSE, col.names = TRUE, ...)
}
\arguments{
\item{x}{The dataframe you want to copy}

\item{row.names}{True if you want row names}

\item{col.names}{False if you don't want column names}

\item{...}{passed to write.table function}
}
\value{
a dataframe in your clipboard, paste it into Excel
}
\description{
Copy dataframe to clipboard for paste into Excel
}
