% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.R
\name{envelope}
\alias{envelope}
\title{Envelope multiple random vibration PSD traces in log space}
\usage{
envelope(freq, PSD, trace, nout = 5000)
}
\arguments{
\item{freq}{array of frequency values}

\item{PSD}{array of PSD values}

\item{trace}{ID (factor levels) for freq/PSD curve}

\item{nout}{number of output pairs, default 5000}
}
\value{
a dataframe with frequency and enveloped PSD level
}
\description{
Envelope multiple random vibration PSD traces in log space
}
