% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBBS.R
\name{GBBS}
\alias{GBBS}
\title{Score Great British Baking Show Guesses}
\usage{
GBBS(result, guess)
}
\arguments{
\item{result}{Result of the show (Vector)}

\item{guess}{Players Guess (Vector)}
}
\value{
Sum of the players Score
}
\description{
Score Great British Baking Show Guesses
}
\examples{
result <- sample(seq(1,10))
guess <- sample(seq(1,10))

score <- GBBS(result,guess)
}
