% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRMS.R
\name{GRMS}
\alias{GRMS}
\title{Calculate GRMS}
\usage{
GRMS(freq, PSD)
}
\arguments{
\item{freq}{PSD frequency in Hz}

\item{PSD}{PSD amplitude in g^2/Hz}
}
\value{
the GRMS value of a PSD
}
\description{
Calculate GRMS
}
