% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_ansys.R
\name{scale_fill_ansys}
\alias{scale_fill_ansys}
\title{ANSYS Color scale used with ggplot graphics}
\usage{
scale_fill_ansys(rev = F, n.breaks = 9, show.limits = T, ...)
}
\arguments{
\item{rev}{number of colors used}

\item{n.breaks}{number of breaks}

\item{show.limits}{Show outer limits of legend}

\item{...}{other arguments to ggplot2::scale_color_stepsn()}
}
\value{
a binned color scale to be used with ggplot graphics.
I recommend using theme(legend.key.height = unit(1, 'in'))) to
adjust the legend height
}
\description{
ANSYS Color scale used with ggplot graphics
}
