% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_x_logTen.R
\name{scale_x_logTen}
\alias{scale_x_logTen}
\title{Define Log 10 X-axis with proper grid settings and labels}
\usage{
scale_x_logTen(...)
}
\arguments{
\item{...}{Arguments to be passed to the scale_x_log10 ggplot2 function}
}
\value{
Log 10 X-axis with proper grid settings and labels
}
\description{
Define Log 10 X-axis with proper grid settings and labels
}
