% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_y_logTen.R
\name{scale_y_logTen}
\alias{scale_y_logTen}
\title{Define Log 10 Y-axis with proper grid settings and labels}
\usage{
scale_y_logTen(...)
}
\arguments{
\item{...}{Arguments to be passed to the scale_y_log10 ggplot2 function}
}
\value{
Log 10 Y-axis with proper grid settings and labels
}
\description{
Define Log 10 Y-axis with proper grid settings and labels
}
