% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_log.R
\name{seq_log}
\alias{seq_log}
\title{Logarithmic sequence}
\usage{
seq_log(from, to, length.out)
}
\arguments{
\item{from}{starting number for the sequence}

\item{to}{ending number for the sequence}

\item{length.out}{number of values in the sequence}
}
\value{
a sequence of length 'length.out', evenly spaced on a logarithmic scale
}
\description{
Logarithmic sequence
}
