% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_swan.R
\name{theme_swan}
\alias{theme_swan}
\title{Theme designed for use with ggplot inspired by hrbrthemes}
\usage{
theme_swan(
  base_family = "IBM Plex Sans",
  base_size = 11.5,
  plot_title_family = base_family,
  plot_title_size = 18,
  plot_title_face = "bold",
  plot_title_margin = 10,
  subtitle_family = base_family,
  subtitle_size = 12,
  subtitle_face = "plain",
  subtitle_margin = 15,
  strip_text_family = base_family,
  strip_text_size = 12,
  strip_text_face = "plain",
  caption_family = base_family,
  caption_size = 9,
  caption_face = "italic",
  caption_margin = 10,
  axis_text_size = base_size,
  axis_title_family = subtitle_family,
  axis_title_size = 10,
  axis_title_face = "plain",
  axis_title_just = "rt",
  plot_margin = margin(30, 30, 30, 30),
  grid_col = "#cccccc",
  grid = TRUE,
  axis_col = "#cccccc",
  axis = TRUE,
  ticks = FALSE,
  caption = "Lockheed Martin Aeronautics",
  legend.key.height = 0.3
)
}
\arguments{
\item{base_family}{base font used in plot}

\item{base_size}{font size}

\item{plot_title_family}{plot title font}

\item{plot_title_size}{plot title size}

\item{plot_title_face}{plot title face}

\item{plot_title_margin}{plot title margin}

\item{subtitle_family}{subtitle font}

\item{subtitle_size}{subtitle font size}

\item{subtitle_face}{subtitle face}

\item{subtitle_margin}{subtitle margin}

\item{strip_text_family}{not really sure}

\item{strip_text_size}{not really sure}

\item{strip_text_face}{not really sure}

\item{caption_family}{caption family}

\item{caption_size}{caption size}

\item{caption_face}{caption face}

\item{caption_margin}{caption margin}

\item{axis_text_size}{axis text size}

\item{axis_title_family}{axis title family}

\item{axis_title_size}{axis title size}

\item{axis_title_face}{axis title face}

\item{axis_title_just}{axis title just}

\item{plot_margin}{plot margin}

\item{grid_col}{grid color}

\item{grid}{grid}

\item{axis_col}{axis color}

\item{axis}{axis}

\item{ticks}{ticks}

\item{caption}{caption for plot, defaults to LM Aeronautics}

\item{legend.key.height}{height of the legend}
}
\value{
a plot theme with nice fonts and colors
}
\description{
This function applies a nice theme to ggplots,
}
